Bugfix: work with perl 5.9.0 or later
Since perl 5.9.0 removed pseudo-hash and introduced restricted-hash,
the behavior of fields and base pragma was entirely changed.
Especially, because restricted-hash is a kind of HASH, if a subclass
and its super-class declares a field with the same name, they share
a single field instead of having a different (private) field for each.
That was the case for this module.
I fixed the issue by renaming a field name avoiding conflict.